feat(inventory,pos): apply CRUD abstractions to Lots/Serials/Users + split POSService
  - Refactor LotForm and LotsView: pure form with formId/onSubmit, useCrudOperations + FormModal
  - Refactor SerialNumberForm and SerialNumbersView: same pattern; StatusChangeDialog unchanged
  - Refactor UsersView: useCrudOperations for createOpen/pagination, ConfirmDialog for deactivation
  - Split POSService.ts (786 lines) into src/services/pos/: POSTypes, ShiftService,
    POSSaleService, POSRefundService, POSReportService, POSProductService, POSCustomerService
  - Update usePOS.ts and all POS view components to import from individual services
  - Delete src/services/POSService.ts
