+++ src/services/pos/POSTypes.ts
@@ -1,0 +1,407 @@
+import type { Meta } from '@/@types/api'
+
+// === ENUMS ===
+export type ShiftStatus = 'OPEN' | 'CLOSED'
+export type POSSaleStatus = 'DRAFT' | 'CONFIRMED' | 'INVOICED' | 'CANCELLED'
+export type POSPaymentStatus = 'PENDING' | 'PARTIAL' | 'PAID'
+export type POSPaymentMethod = 'CASH' | 'CARD' | 'TRANSFER' | 'CREDIT'
+export type CashMovementType = 'IN' | 'OUT'
+export type RefundStatus = 'PENDING' | 'COMPLETED' | 'CANCELLED'
+export type DiscountType = 'PERCENTAGE' | 'AMOUNT'
+
+// === LABELS ===
+export const SHIFT_STATUS_LABELS: Record<ShiftStatus, string> = {
+    OPEN: 'Abierto',
+    CLOSED: 'Cerrado',
+}
+
+export const SHIFT_STATUS_CLASSES: Record<ShiftStatus, string> = {
+    OPEN: 'bg-emerald-100 text-emerald-700 dark:bg-emerald-500/20 dark:text-emerald-300',
+    CLOSED: 'bg-gray-100 text-gray-700 dark:bg-gray-500/20 dark:text-gray-300',
+}
+
+export const POS_SALE_STATUS_LABELS: Record<POSSaleStatus, string> = {
+    DRAFT: 'Borrador',
+    CONFIRMED: 'Confirmada',
+    INVOICED: 'Facturada',
+    CANCELLED: 'Cancelada',
+}
+
+export const POS_SALE_STATUS_CLASSES: Record<POSSaleStatus, string> = {
+    DRAFT: 'bg-gray-100 text-gray-700 dark:bg-gray-500/20 dark:text-gray-300',
+    CONFIRMED:
+        'bg-blue-100 text-blue-700 dark:bg-blue-500/20 dark:text-blue-300',
+    INVOICED:
+        'bg-emerald-100 text-emerald-700 dark:bg-emerald-500/20 dark:text-emerald-300',
+    CANCELLED: 'bg-red-100 text-red-700 dark:bg-red-500/20 dark:text-red-300',
+}
+
+export const POS_PAYMENT_STATUS_LABELS: Record<POSPaymentStatus, string> = {
+    PENDING: 'Pendiente',
+    PARTIAL: 'Parcial',
+    PAID: 'Pagado',
+}
+
+export const POS_PAYMENT_STATUS_CLASSES: Record<POSPaymentStatus, string> = {
+    PENDING:
+        'bg-amber-100 text-amber-700 dark:bg-amber-500/20 dark:text-amber-300',
+    PARTIAL: 'bg-blue-100 text-blue-700 dark:bg-blue-500/20 dark:text-blue-300',
+    PAID: 'bg-emerald-100 text-emerald-700 dark:bg-emerald-500/20 dark:text-emerald-300',
+}
+
+export const POS_PAYMENT_METHOD_LABELS: Record<POSPaymentMethod, string> = {
+    CASH: 'Efectivo',
+    CARD: 'Tarjeta',
+    TRANSFER: 'Transferencia',
+    CREDIT: 'Crédito',
+}
+
+export const CASH_MOVEMENT_TYPE_LABELS: Record<CashMovementType, string> = {
+    IN: 'Ingreso',
+    OUT: 'Egreso',
+}
+
+export const CASH_MOVEMENT_TYPE_CLASSES: Record<CashMovementType, string> = {
+    IN: 'bg-emerald-100 text-emerald-700 dark:bg-emerald-500/20 dark:text-emerald-300',
+    OUT: 'bg-red-100 text-red-700 dark:bg-red-500/20 dark:text-red-300',
+}
+
+export const REFUND_STATUS_LABELS: Record<RefundStatus, string> = {
+    PENDING: 'Pendiente',
+    COMPLETED: 'Completado',
+    CANCELLED: 'Cancelado',
+}
+
+export const REFUND_STATUS_CLASSES: Record<RefundStatus, string> = {
+    PENDING:
+        'bg-amber-100 text-amber-700 dark:bg-amber-500/20 dark:text-amber-300',
+    COMPLETED:
+        'bg-emerald-100 text-emerald-700 dark:bg-emerald-500/20 dark:text-emerald-300',
+    CANCELLED: 'bg-red-100 text-red-700 dark:bg-red-500/20 dark:text-red-300',
+}
+
+// === SHIFTS ===
+export interface Shift {
+    id: number
+    cashierName: string
+    openedAt: string
+    closedAt: string | null
+    openingBalance: number
+    closingBalance: number | null
+    expectedBalance: number | null
+    discrepancy: number | null
+    status: ShiftStatus
+    notes: string | null
+}
+
+export interface OpenShiftInput {
+    cashierName: string
+    openingBalance: number
+    notes?: string
+}
+
+export interface CloseShiftInput {
+    closingBalance: number
+    notes?: string
+}
+
+// === SALES ===
+export interface POSSale {
+    id: number
+    customerId: number | null
+    isFinalConsumer: boolean
+    shiftId: number
+    status: POSSaleStatus
+    saleDate: string | null
+    subtotal: number
+    tax: number
+    discount: number
+    discountType: DiscountType | null
+    discountValue: number
+    total: number
+    paymentStatus: POSPaymentStatus
+    notes: string | null
+    createdBy: string | null
+    parkedAt: string | null
+    parkReason: string | null
+    createdAt: string
+}
+
+export interface POSSaleItem {
+    id: number
+    saleId: number
+    productId: number
+    quantity: number
+    unitPrice: number
+    discount: number
+    taxRate: number
+    taxAmount: number
+    subtotal: number
+    priceOverride: number | null
+    overrideReason: string | null
+}
+
+export interface POSPayment {
+    id: number
+    saleId: number
+    amount: number
+    paymentMethod: POSPaymentMethod
+    paymentDate: string
+    reference: string | null
+    notes: string | null
+}
+
+// === INPUTS ===
+export interface CreateSaleInput {
+    customerId?: number | null
+    isFinalConsumer?: boolean
+    notes?: string
+    createdBy?: string
+}
+
+export interface QuickSaleInput {
+    customerId?: number | null
+    items: QuickSaleItemInput[]
+    payments: QuickSalePaymentInput[]
+    discountType?: DiscountType
+    discountValue?: number
+    notes?: string
+    createdBy?: string
+}
+
+export interface QuickSaleItemInput {
+    productId: number
+    quantity: number
+    unitPrice?: number
+    discount?: number
+}
+
+export interface QuickSalePaymentInput {
+    amount: number
+    paymentMethod: POSPaymentMethod
+    reference?: string
+    notes?: string
+}
+
+export interface AddSaleItemInput {
+    productId: number
+    quantity: number
+    unitPrice: number
+    discount?: number
+}
+
+export interface UpdateSaleItemInput {
+    quantity?: number
+    discount?: number
+}
+
+export interface SaleDiscountInput {
+    discountType: DiscountType
+    discountValue: number
+}
+
+export interface PriceOverrideInput {
+    newPrice: number
+    reason: string
+}
+
+export interface PaymentInput {
+    amount: number
+    paymentMethod: POSPaymentMethod
+    reference?: string
+    notes?: string
+}
+
+// === REFUNDS ===
+export interface Refund {
+    id: number
+    originalSaleId: number
+    shiftId: number
+    refundDate: string | null
+    subtotal: number
+    tax: number
+    total: number
+    reason: string | null
+    status: RefundStatus
+    refundedBy: string | null
+}
+
+export interface RefundItem {
+    id: number
+    refundId: number
+    originalSaleItemId: number
+    productId: number
+    quantity: number
+    unitPrice: number
+    discount: number
+    taxRate: number
+    taxAmount: number
+    subtotal: number
+}
+
+export interface CreateRefundInput {
+    originalSaleId: number
+    items: { saleItemId: number; quantity: number }[]
+    reason?: string
+    refundedBy?: string
+}
+
+export interface ProcessRefundInput {
+    payments: {
+        amount: number
+        paymentMethod: POSPaymentMethod
+        reference?: string
+    }[]
+}
+
+// === CASH MOVEMENTS ===
+export interface CashMovement {
+    id: number
+    shiftId: number
+    type: CashMovementType
+    amount: number
+    reason: string | null
+    performedBy: string | null
+    createdAt: string
+}
+
+export interface CashMovementInput {
+    type: CashMovementType
+    amount: number
+    reason?: string
+    performedBy?: string
+}
+
+export interface CashSummary {
+    shiftId: number
+    openingBalance: number
+    cashSales: number
+    cashRefunds: number
+    cashIn: number
+    cashOut: number
+    expectedBalance: number
+}
+
+// === RECEIPT ===
+export interface ReceiptItem {
+    productName: string
+    quantity: number
+    unitPrice: number
+    discount: number
+    discountAmount: number
+    taxRate: number
+    taxAmount: number
+    subtotal: number
+    priceOverride: number | null
+    overrideReason: string | null
+}
+
+export interface Receipt {
+    saleId: number
+    saleDate: string
+    status: string
+    cashier: string
+    customer: { name: string; taxId: string } | null
+    isFinalConsumer: boolean
+    items: ReceiptItem[]
+    taxBreakdown: {
+        taxRate: number
+        taxableBase: number
+        taxAmount: number
+    }[]
+    subtotal: number
+    discount: number
+    discountType: DiscountType | null
+    discountValue: number
+    tax: number
+    total: number
+    payments: {
+        method: string
+        amount: number
+        reference: string | null
+    }[]
+    totalPaid: number
+    change: number
+}
+
+// === REPORTS ===
+export interface XReport {
+    shift: Shift
+    salesSummary: {
+        count: number
+        subtotal: number
+        tax: number
+        discount: number
+        total: number
+    }
+    paymentsByMethod: {
+        paymentMethod: string
+        count: number
+        total: number
+    }[]
+    itemsSold: {
+        productName: string
+        sku: string
+        quantity: number
+        total: number
+    }[]
+}
+
+export interface ZReport extends XReport {
+    refundSummary: { count: number; total: number }
+    cashReconciliation: CashSummary & {
+        closingBalance: number
+        discrepancy: number
+    }
+}
+
+export interface DailyReport {
+    date: string
+    totalSales: number
+    totalAmount: number
+    paymentsByMethod: {
+        paymentMethod: string
+        count: number
+        total: number
+    }[]
+    topProducts: {
+        productName: string
+        sku: string
+        quantity: number
+        total: number
+    }[]
+    refundSummary: { count: number; total: number }
+}
+
+// === POS PRODUCT ===
+export interface POSProduct {
+    id: number
+    name: string
+    sku: string
+    barcode: string | null
+    categoryId: number | null
+    salePrice: number | null
+    taxRate: number
+    isActive: boolean
+    image: string | null
+}
+
+// === POS CUSTOMER ===
+export interface POSCustomer {
+    id: number
+    name: string
+    taxId: string
+    taxType: string
+}
+
+export interface QuickCustomerInput {
+    name: string
+    taxId: string
+    taxType: string
+}
+
+// === SHARED ===
+export interface ListResponse<T> {
+    data: T[]
+    meta: Meta
+}
